Add toogleComment feature on Article (Comment'deletedAt' is null or not)

+++ www/Models/Article.php
@@ -441,24 +441,48 @@ class Article extends Database implements JsonSerializable
 
         $this->setPublicationDate($today);
         $this->setDeletedAt(null);
+
+        if (!empty($this->getId())) {
+            $this->toggleComments();
+        }
     }
 
     public function setToScheduled($publicationDate) {
         $this->setPublicationDate(htmlspecialchars($publicationDate));
         $this->setDeletedAt(null);
+
+        if (!empty($this->getId())) {
+            $this->toggleComments();
+        }
     }
 
     public function setToDraft() {
         $this->setPublicationDate(null);
         $this->setDeletedAt(null);
+
+        if (!empty($this->getId())) {
+            $this->toggleComments();
+        }
     }
 
     public function articleSoftDelete() {
+        $this->toggleComments("hide");
+        $this->delete();
+    }
+
+    // state == display / hide
+    public function toggleComments($state = "display") {
+        if (!($state == "display" || $state == "hide")) return;
+
         $comments = $this->getComments();
         foreach ($comments as $comment) {
-            $comment->delete();
+            if ($state == "hide") {
+                $comment->delete();
+            } else {
+                $comment->setDeletedAt(null);
+                $comment->save();
+            }
         }
-        $this->delete();
     }
 
     public function articleHardDelete() {
